+++ ProjectFeedget/.idea/modules.xml
@@ -3,7 +3,6 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/ProjectFeedget.iml" filepath="$PROJECT_DIR$/ProjectFeedget.iml" />
-      <module fileurl="file://$PROJECT_DIR$/ProjectNoname.iml" filepath="$PROJECT_DIR$/ProjectNoname.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
