+++ ProjectFeedget/app/src/main/java/kr/mashup/feedget/MyPageActivity.java
@@ -1,84 +1,13 @@
 package kr.mashup.feedget;
 
 import android.os.Bundle;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-import android.widget.ImageButton;
-
-import kr.mashup.feedget.ui.main.tabs.creation.CreationFragment;
 
 public class MyPageActivity extends AppCompatActivity {
-
-    Fragment[] arrFragments;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_my_page);
-
-        ViewPager viewPager = findViewById(R.id.vp_pager);
-        TabLayout tabLayout = findViewById(R.id.tl_tabs);
-
-        arrFragments = new Fragment[2];
-        arrFragments[0] = new CreationFragment();
-        arrFragments[1] = new CreationFragment();
-
-        MainPagerAdapter adapter = new MainPagerAdapter(getSupportFragmentManager(), arrFragments);
-
-        viewPager.setAdapter(adapter);
-        tabLayout.setupWithViewPager(viewPager);
-
-        ImageButton backButton = findViewById(R.id.backButton);
-        backButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                finish();
-            }
-        });
-
     }
-
-    class MainPagerAdapter extends FragmentPagerAdapter {
-
-        Fragment[] arrFragments;
-
-        public MainPagerAdapter(FragmentManager fm, Fragment[] arrFragments) {
-            super(fm);
-            this.arrFragments = arrFragments;
-        }
-
-
-
-
-        @Override
-        public CharSequence getPageTitle(int position) {
-
-            switch (position) {
-                case 0:
-                    return "게시글";
-                case 1:
-                    return "남긴 피드백";
-
-
-                default:
-                    return "";
-            }
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            return arrFragments[position];
-        }
-
-        @Override
-        public int getCount() {
-            return arrFragments.length;
-        }
-    }
-
 }
