+++ ProjectFeedget/app/src/main/res/layout/activity_my_page.xml
@@ -3,187 +3,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
     tools:context="kr.mashup.feedget.MyPageActivity">
 
-    <LinearLayout
-        android:background="#ffffff"
-        android:orientation="vertical"
+    <fragment
         android:layout_width="match_parent"
-        android:layout_height="312dp">
+        android:layout_height="match_parent"
+        android:id="@+id/fragment_my_page"
+        android:name="kr.mashup.feedget.MyPageFragment" />
 
-        <FrameLayout
-            android:background="#ffffff"
-            android:gravity="center_vertical"
-            android:orientation="horizontal"
-            android:layout_width="match_parent"
-            android:layout_height="48dp">
-
-            <ImageButton
-                android:id="@+id/backButton"
-                android:background="@android:color/transparent"
-                android:src="@drawable/ic_back"
-                android:layout_width="48dp"
-                android:layout_height="48dp" />
-
-            <TextView
-                android:textSize="16sp"
-                android:textColor="#000000"
-                android:layout_gravity="center_vertical"
-                android:layout_marginLeft="56dp"
-                android:text="마이페이지"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-        </FrameLayout>
-        <View
-            android:background="#ffffff"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"/>
-
-        <ImageView
-            android:layout_marginTop="20dp"
-            android:layout_gravity="center"
-            android:background="@drawable/active_dot"
-            android:layout_width="100dp"
-            android:layout_height="100dp" />
-
-
-        <LinearLayout
-            android:layout_marginTop="20dp"
-            android:gravity="center_horizontal"
-            android:orientation="horizontal"
-            android:layout_width="match_parent"
-            android:layout_height="20dp">
-
-            <TextView
-                android:text="닉네임"
-                android:textSize="16sp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="24dp"
-            android:orientation="horizontal"
-            android:gravity="center_horizontal">
-
-        <LinearLayout
-            android:gravity="center_horizontal"
-            android:orientation="vertical"
-            android:layout_width="94dp"
-            android:layout_height="50dp">
-
-            <TextView
-                android:text="채택률"
-                android:gravity="center_horizontal"
-                android:textSize="12sp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-            <TextView
-                android:text="100%"
-                android:layout_marginTop="4dp"
-                android:gravity="center_horizontal"
-                android:textSize="14sp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:gravity="center_horizontal"
-            android:layout_marginLeft="16dp"
-            android:orientation="vertical"
-            android:layout_width="94dp"
-            android:layout_height="50dp">
-
-            <TextView
-                android:text="보유 포인트"
-                android:gravity="center_horizontal"
-                android:textSize="12sp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-            <TextView
-                android:text="99,999,999"
-                android:layout_marginTop="4dp"
-                android:gravity="center_horizontal"
-                android:textSize="14sp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:gravity="center_horizontal"
-            android:layout_marginLeft="16dp"
-            android:orientation="vertical"
-            android:layout_width="94dp"
-            android:layout_height="50dp">
-
-            <TextView
-                android:text="등급"
-                android:gravity="center_horizontal"
-                android:textSize="12sp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-            <TextView
-                android:text="등급이름"
-                android:layout_marginTop="4dp"
-                android:gravity="center_horizontal"
-                android:textSize="14sp"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content" />
-
-        </LinearLayout>
-
-        </LinearLayout>
-
-    </LinearLayout>
-
-    <android.support.design.widget.TabLayout
-        android:id="@+id/tl_tabs"
-        app:tabIndicatorColor="#000000"
-        app:tabIndicatorHeight="3dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
-    
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:gravity="center"
-        android:layout_height="40dp"
-        android:orientation="horizontal">
-
-        <TextView
-            android:layout_width="164dp"
-            android:layout_height="18dp"
-            android:textSize="12sp"
-            android:text="(3/3)"/>
-
-        <TextView
-            android:layout_width="62dp"
-            android:layout_height="18dp"
-            android:layout_marginLeft="75dp"
-            android:textSize="12sp"
-            android:text="전체"/>
-
-        <ImageView
-            android:background="@drawable/ic_expand_more_black_24dp"
-            android:layout_width="24dp"
-            android:layout_height="24dp" />
-
-
-    </LinearLayout>
-
-    <android.support.v4.view.ViewPager
-        android:layout_below="@id/tl_tabs"
-        android:id="@+id/vp_pager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
 </LinearLayout>
